% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpatialExperiment.R
\name{loadSpatialExperiment}
\alias{loadSpatialExperiment}
\title{Load a spatial experiment}
\usage{
loadSpatialExperiment(exp.info, project)
}
\arguments{
\item{exp.info}{Named list of metadata for a spatial 'omics experiment.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link[alabaster.base]{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{SpatialExperiment} object.
}
\description{
Load a \linkS4class{SpatialExperiment} object from its constituent files in DataSetDB.
}
\examples{
library(SpatialExperiment)
example(read10xVisium, echo=FALSE)
colnames(spe) <- make.unique(colnames(spe)) # forcing unique column names.

tmp <- tempfile()
dir.create(tmp)
meta <- stageObject(spe, tmp, "experiment-1")

meta$path <- "experiment-1/experiment.json"
loadSpatialExperiment(meta, tmp)

}
\author{
Aaron Lun
}
