% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageSpatialExperiment.R
\name{stageObject,SpatialExperiment-method}
\alias{stageObject,SpatialExperiment-method}
\title{Stage a spatial experiment}
\usage{
\S4method{stageObject}{SpatialExperiment}(x, dir, path, child = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{SpatialExperiment} object.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{...}{Further arguments to pass to specific methods.}
}
\value{
A named list of the same form as that returned by the \code{\link{stageObject}} method for a SingleCellExperiment,
but containing additional fields for the spatial data.
A directory is created at \code{path} inside \code{dir} and is populated with the contents of \code{x}.
}
\description{
Stage a \linkS4class{SpatialExperiment} object.
}
\examples{
library(SpatialExperiment)
example(read10xVisium, echo=FALSE)
colnames(spe) <- make.unique(colnames(spe)) # forcing unique column names.

tmp <- tempfile()
dir.create(tmp)
stageObject(spe, tmp, "experiment-1")
list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
